+++ lib/libvector/v3_dot.c
@@ -12,7 +12,7 @@
 
 #include "libvector.h"
 
-float	v3_dot(t_vector3 v1, t_vector3 v2)
+double	v3_dot(t_vector3 v1, t_vector3 v2)
 {
 	return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z);
 }
