+++ src/simulate/hook_key_event.c
@@ -17,7 +17,7 @@ void	move_camera(void *param)
 {
 	t_world *const		world = param;
 	t_element *const	camera = &(world->camera->base);
-	const float			delta = 0.1;
+	const double		delta = 0.1;
 	t_vector3			v[3];
 
 	enum e_type {FRONT, RIGHT, NEW};
@@ -43,7 +43,7 @@ void	transform_objs(void *param)
 {
 	t_world *const		w = param;
 	t_element *const	c = &(w->camera->base);
-	const t_vector3		delta = v3_mul(v3_preset(V3_ONE), 0.1f);
+	const t_vector3		delta = v3_mul(v3_preset(V3_ONE), 0.1);
 
 	if (w->selected && w->gui.keyboard[KEYBOARD_p])
 		w->transform_type = Position;
